collection-classification/requirements.txt
appdirs==1.4.0
click==6.6
cupy==1.0.0a1
Cython==0.25.2
decorator==4.0.10
filelock==2.0.7
Flask==0.12
h5py==2.6.0
itsdangerous==0.24
Jinja2==2.8.1
Keras==1.2.1
MarkupSafe==0.23
nose==1.3.7
numpy==1.11.3
Pillow==3.4.2
protobuf==3.1.0
py==1.4.32
pydot-ng==1.0.0
pyparsing==2.1.10
pytest==3.0.5
pytools==2016.2.6
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow-gpu==0.12.0
Theano==0.8.2
Werkzeug==0.11.15
